Removing the license section
With the merge of the org, there are repos that are under Facebook BSD license which is not MIT/ISC nor owned by a non-profit (Facebook). I believe that "the owners of a project can continue to maintain it" is mostly true, regardless of the license of acquisitions. The problem usually come when the owners leave and the project is left to an uncertain future.